+++ src/Sulu/Bundle/WebsiteBundle/Controller/CacheController.php
@@ -10,6 +10,8 @@
 
 namespace Sulu\Bundle\WebsiteBundle\Controller;
 
+use Sulu\Bundle\ContentBundle\Admin\ContentAdmin;
+use Sulu\Component\Security\Authorization\PermissionTypes;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\JsonResponse;
 
@@ -25,8 +27,32 @@ class CacheController extends Controller
      */
     public function clearAction()
     {
+        if (!$this->checkLivePermissionForAllWebspaces()) {
+            return new JsonResponse(null, 403);
+        }
+
         $this->get('sulu_website.http_cache.clearer')->clear();
 
         return new JsonResponse([], 200);
+    }
+
+    /**
+     * Check the permissions for all webspaces.
+     * Returns true if the user has live permission in all webspaces.
+     *
+     * TODO should be replaced with a single webspace cache clear.
+     *
+     * @return bool
+     */
+    private function checkLivePermissionForAllWebspaces()
+    {
+        foreach ($this->get('sulu_core.webspace.webspace_manager')->getWebspaceCollection() as $webspace) {
+            $context = ContentAdmin::SECURITY_CONTEXT_PREFIX . $webspace->getKey();
+            if (!$this->get('sulu_security.security_checker')->hasPermission($context, PermissionTypes::LIVE)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 }